+++ src/Network/Network/Properties/Resources.Designer.cs
@@ -464,7 +464,18 @@ namespace Microsoft.Azure.Commands.Network.Properties {
                 return ResourceManager.GetString("CreatingLongRunningOperationMessage", resourceCulture);
             }
         }
-        
+
+        /// <summary>
+        ///   Looks up a localized string similar to Creating child resource with ResourceGroupName {0}, ParentResourceName {1} ResourceName {2}..
+        /// </summary>
+        internal static string CreatingChildResourceLongRunningOperationMessage
+        {
+            get
+            {
+                return ResourceManager.GetString("CreatingChildResourceLongRunningOperationMessage", resourceCulture);
+            }
+        }
+
         /// <summary>
         ///   Looks up a localized string similar to Creating Resource.
         /// </summary>
@@ -1067,7 +1078,18 @@ namespace Microsoft.Azure.Commands.Network.Properties {
                 return ResourceManager.GetString("ParentVpnGatewayNotFound", resourceCulture);
             }
         }
-        
+
+        /// <summary>
+        ///   Looks up a localized string similar to The parent VpnServerConfiguration for this PolicyGroup cannot be found..
+        /// </summary>
+        internal static string ParentVpnServerConfigurationNotFound
+        {
+            get
+            {
+                return ResourceManager.GetString("ParentVpnServerConfigurationNotFound", resourceCulture);
+            }
+        }
+
         /// <summary>
         ///   Looks up a localized string similar to IP version is undefined.
         /// </summary>
@@ -1742,7 +1764,18 @@ namespace Microsoft.Azure.Commands.Network.Properties {
                 return ResourceManager.GetString("VpnConnectionNotFound", resourceCulture);
             }
         }
-        
+
+        /// <summary>
+        ///   Looks up a localized string similar to The VpnServerConfigurationPolicyGroup could not be found..
+        /// </summary>
+        internal static string VpnServerConfigurationPolicyGroupNotFound
+        {
+            get
+            {
+                return ResourceManager.GetString("VpnServerConfigurationPolicyGroupNotFound", resourceCulture);
+            }
+        }
+
         /// <summary>
         ///   Looks up a localized string similar to Update request for VpnConnection contains deprecated property. Please use corresponding VpnSiteLinkConnection property instead..
         /// </summary>
@@ -1778,7 +1811,18 @@ namespace Microsoft.Azure.Commands.Network.Properties {
                 return ResourceManager.GetString("VpnGatewayRequiredToCreateVpnConnection", resourceCulture);
             }
         }
-        
+
+        /// <summary>
+        ///   Looks up a localized string similar to A valid VpnServerConfiguration reference is required to create or update a PolicyGroup..
+        /// </summary>
+        internal static string VpnServerConfigurationRequiredToCreateOrUpdatePolicyGroup
+        {
+            get
+            {
+                return ResourceManager.GetString("VpnServerConfigurationRequiredToCreateOrUpdatePolicyGroup", resourceCulture);
+            }
+        }
+
         /// <summary>
         ///   Looks up a localized string similar to A valid VpnGateway reference is required to create a VpnNatRule.
         /// </summary>
@@ -1805,7 +1849,7 @@ namespace Microsoft.Azure.Commands.Network.Properties {
                 return ResourceManager.GetString("VpnServerConfigurationNotFound", resourceCulture);
             }
         }
-        
+
         /// <summary>
         ///   Looks up a localized string similar to A valid VpnServerConfiguration is required to create a P2SVpnGateway.
         /// </summary>